Create style.css
Broadway
In this project, you will use properties such as display and position to improve the layout of the landing page for a fictional design firm, Broadway Design.

The site has some style rules to begin with. You will improve the layout and positioning of the navigation menu at the top of the page and the three supporting sections (Design, Develop, Deploy) below the image.

If you get stuck during this project or would like to see an experienced developer work through it, click “Get Unstuck“ to see a project walkthrough video.

Tasks
10/10 Complete
Mark the tasks as complete by checking them off
1.
The <header> currently scrolls with the rest of the document.

Set its position property so that it stays stuck to the top of the window when the document is scrolled.


Stuck? Get a hint
2.
The <header> has shrunk!

Change the width of the same element so that it stretches across its entire parent element.


Stuck? Get a hint
3.
List items (<li>) inside of the navigation section (<nav>) are currently displayed as a list.

Set their display property so that they can appear next to each other horizontally (but so that you still set their width in the next task).


Stuck? Get a hint
4.
Set the width of the same elements to 80 pixels.

5.
After changing the position of the <header> element to fixed, the contents of the entire site after it shifted upwards.

Set the position of <main> so that we can position it relatively.


Stuck? Get a hint
6.
The <header> has disappeared behind the <main>.

Use z-index to make the <header> visible.


Stuck? Get a hint
7.
Now the navigation bar looks good, but it is blocking the title at the top of the page.

Offset <main> by 80 pixels from the top.


Stuck? Get a hint
8.
Now, fix up the supporting element style below the image.

Add a declaration to the .supporting .col rule set so that these elements can appear horizontally next to each other but have defined height and width.


Stuck? Get a hint
9.
Inspect the .supporting .col elements—they don’t seem to be flowing horizontally yet because they have no set width. Set the width and height of .supporting .col elements to 200 pixels.


Stuck? Get a hint
10.
Great work, the Broadway Design site looks much better!

If you want to continue coding, challenge yourself to make the <footer> element also fixed to the bottom of the page regardless of scrolling.